Fix missing points issue
What's up
When a user signs up, Github sends us the user's token to authenticate with Github on their behalf. We also generate a separate access token to authenticate with the frontend client. Currently, we are throwing away the Github token and trying to authenticate with Github using our own custom generated token.

What this does
- Persists the Github-provided token as `api_token` in the `users` table.
- Fixes the github client index to be zero-based.