Use active users to generate github client
# What's up
Github's API only allows limited operations with application authorization. To do more advanced things like read comments on a pull request, an application will have to be authorized by a user to use his/her access tokens to do those actions. Currently, we use a predetermined list of users stored in an environment variable `USERS` as the defacto users whose authorizations will be used to carry out these kinds of operations. A problem arises when one or more of those predetermined users no longer authorize the application. To solve this, we should move away from using the `USERS` variable and instead query for active access tokens which is guaranteed to be users who have most recently authorized the application.