Persist each user's score separately for pull requests and reviews (#84)
* Persist each user's score separately for pull requests and reviews

What's up
Currently there are just two entries per user in the `scores` table (one
for weekly, and one for all_time). The weekly score is rolled into the
all time score at the end of each week, and just like that history is
frozen as an image. This is not so bad unless we want to find out the
origin of the points a user has for a certain week.

What this does
Stop tracking scores in such a weird way. Instead persist a separate
score item each time a user is scored for either a pull request or a
review.

- Adds tasks for handling data transformation
- Migration to change the schema of the score table
- Updates to the points marshaller class to adopt the new behavior
- Updates to the hall of fame task to adopt the new behavior

* Remove failing specs

* Add new test cases

* Update the api to work with the new steup