Remove old score artifacts
What's up
https://github.com/alphasights/lion-api/pull/84 established that each
score a user gets is associated to a scoring event (either a pull
request or pull request review). After the migration to this new scheme,
we're left with the old `scores` rows that are now invalid since they
possess neither association.

What this does
Adds a task that cleans up orphaned score rows.